+++ src/main/java/com/loha/flippedclassroom/controller/TeacherPcController.java
@@ -28,34 +28,34 @@ public class TeacherPcController {
 
     @GetMapping(value = "/index")
     public String index(Model model) throws Exception{
-        Integer id=teacherService.getCurTeacher().getId();
+        Long id=teacherService.getCurTeacher().getId();
         model.addAttribute("curTeacherId",id);
         model.addAttribute("courseList",teacherService.getTeacherCourses(id));
         return "pc/TeacherIndex";
     }
 
     @PostMapping(value = "/course")
-    public String courseInfo(Integer courseId,Model model){
+    public String courseInfo(Long courseId,Model model){
         model.addAttribute("courseId",courseId);
         return "pc/coursePage";
     }
 
     @GetMapping(value = "/course/importStudent")
-    public String importStudentPage(Integer courseId,Model model) throws Exception{
+    public String importStudentPage(Long courseId,Model model) throws Exception{
         model.addAttribute("courseName",teacherService.getCourseById(courseId).getCourseName());
         model.addAttribute("klassList",teacherService.getKlassByCourseId(courseId));
         return "pc/importStudent";
     }
 
     @GetMapping(value = "/course/seminar")
-    public String chooseSeminar(Integer courseId,Model model) throws Exception{
+    public String chooseSeminar(Long courseId,Model model) throws Exception{
         model.addAttribute("courseName",teacherService.getCourseById(courseId).getCourseName());
         model.addAttribute("roundAndSeminarList",teacherService.getRoundAndSeminar(courseId));
         return "pc/seminarPage";
     }
 
     @GetMapping(value = "/course/exportScore")
-    public String exportScore(Integer courseId,Model model) throws Exception{
+    public String exportScore(Long courseId,Model model) throws Exception{
         model.addAttribute("courseName",teacherService.getCourseById(courseId).getCourseName());
         return "pc/exportScore";
     }
