+++ src/main/java/com/loha/flippedclassroom/entity/SeminarScore.java
@@ -13,8 +13,8 @@ import java.math.BigDecimal;
 @Getter
 @Setter
 public class SeminarScore {
-    private Integer klassSeminarId;
-    private Integer teamId;
+    private Long klassSeminarId;
+    private Long teamId;
     private BigDecimal totalScore;
     private BigDecimal presentationScore;
     private BigDecimal questionScore;
